refactor: lifecycle 관련 dto를 별도 패키지 안으로 이동
