+++ src/main/java/umc/cockple/demo/domain/exercise/dto/lifecycle/ExerciseCreateDTO.java
@@ -1,4 +1,4 @@
-package umc.cockple.demo.domain.exercise.dto;
+package umc.cockple.demo.domain.exercise.dto.lifecycle;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.*;
